requirements.txt
droidrun>=0.2.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
